conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=py38h06a4308_0
aiohttp=3.7.4=py38h27cfd23_1
argon2-cffi=20.1.0=py38h27cfd23_1
async-timeout=3.0.1=py38h06a4308_0
async_generator=1.10=pyhd3eb1b0_0
attrs=20.3.0=pyhd3eb1b0_0
autopep8=1.5.6=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
bleach=3.3.0=pyhd3eb1b0_0
blinker=1.4=py38h06a4308_0
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.1.19=h06a4308_1
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2020.12.5=py38h06a4308_0
cffi=1.14.5=py38h261ae71_0
chardet=3.0.4=py38h06a4308_1003
click=7.1.2=pyhd3eb1b0_0
coverage=5.5=py38h27cfd23_2
cryptography=3.4.7=py38hd23ed53_0
cudatoolkit=10.1.243=h6bb024c_0
cython=0.29.22=py38h2531618_0
dbus=1.13.18=hb2f20db_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.3=py38_0
expat=2.3.0=h2531618_2
ffmpeg=4.2.2=h20bf706_0
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
fsspec=0.8.7=pyhd3eb1b0_0
future=0.18.2=py38_1
glib=2.68.0=h36276a3_0
gmp=6.2.1=h2531618_2
gnutls=3.6.5=h71b1129_1002
google-auth=1.28.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.3=pyhd3eb1b0_0
grpcio=1.36.1=py38h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.7.3=py38h06a4308_1
importlib_metadata=3.7.3=hd3eb1b0_1
intel-openmp=2020.2=254
ipykernel=5.3.4=py38h5ca1d4c_0
ipython=7.22.0=py38hd0cf306_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3eb1b0_1
jedi=0.18.0=py38h06a4308_1
jinja2=2.11.3=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
jsonnet=0.17.0=py38hadf7658_0
jsonschema=3.2.0=py_2
jupyter=1.0.0=py38_7
jupyter_client=6.1.7=py_0
jupyter_console=6.3.0=pyhd3eb1b0_0
jupyter_core=4.7.1=py38h06a4308_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
lame=3.100=h7b6447c_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=h1bed415_2
libuv=1.40.0=h7b6447c_0
libvpx=1.7.0=h439df22_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py38h06a4308_0
markupsafe=1.1.1=py38h7b6447c_0
mistune=0.8.4=py38h7b6447c_1000
mkl=2020.2=256
mkl-service=2.3.0=py38he904b0f_0
mkl_fft=1.3.0=py38h54f3939_0
mkl_random=1.1.1=py38h0573a6f_0
multidict=5.1.0=py38h27cfd23_2
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.0.7=py38_0
nbformat=5.1.2=pyhd3eb1b0_1
ncurses=6.2=he6710b0_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.4.1=hbb512f6_0
ninja=1.10.2=py38hff7bd54_0
notebook=6.3.0=py38h06a4308_0
numpy=1.19.2=py38h54aff64_0
numpy-base=1.19.2=py38hfa32c7d_0
oauthlib=3.1.0=py_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
packaging=20.9=pyhd3eb1b0_0
pandoc=2.12=h06a4308_0
pandocfilters=1.4.3=py38h06a4308_1
parso=0.8.1=pyhd3eb1b0_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.2=py38he98fc37_0
pip=21.0.1=py38h06a4308_0
prometheus_client=0.9.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
prompt_toolkit=3.0.17=hd3eb1b0_0
protobuf=3.14.0=py38h2531618_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycodestyle=2.7.0=pypi_0
pycparser=2.20=py_2
pygments=2.8.1=pyhd3eb1b0_0
pyjwt=2.0.1=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py38h05f1152_4
pyrsistent=0.17.3=py38h7b6447c_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.8=hdb3f193_4
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.8=1_cp38
pytorch=1.8.1=py3.8_cuda10.1_cudnn7.6.3_0
pytorch-lightning=1.2.5=pyhd8ed1ab_0
pyyaml=5.4.1=py38h27cfd23_1
pyzmq=20.0.0=py38h2531618_1
qt=5.9.7=h5867ecd_1
qtconsole=5.0.3=pyhd3eb1b0_0
qtpy=1.9.0=py_0
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py38h06a4308_0
sip=4.19.13=py38he6710b0_0
six=1.15.0=py38h06a4308_0
sqlite=3.35.2=hdfb4753_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
terminado=0.9.3=py38h06a4308_0
testpath=0.4.4=pyhd3eb1b0_0
tk=8.6.10=hbc83047_0
torchaudio=0.8.1=py38
torchmetrics=0.2.0=pyhd8ed1ab_0
torchvision=0.9.1=py38_cu101
tornado=6.1=py38h27cfd23_0
tqdm=4.59.0=pyhd3eb1b0_1
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
urllib3=1.26.4=pyhd3eb1b0_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py38_1
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py38_0
x264=1!157.20191217=h7b6447c_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py38h27cfd23_0
zeromq=4.3.4=h2531618_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
